+++ browser/index.js
@@ -1,8 +1,8 @@
 export { default as default } from '../modules/export/readXlsxFileBrowser.js'
 export { default as readSheet } from '../modules/export/readSheetBrowser.js'
 
-// `parseData()`
-export { default as parseData } from '../modules/parseData/parseData.js'
-export { default as Integer } from '../modules/parseData/types/additional/Integer.js'
-export { default as Email } from '../modules/parseData/types/additional/Email.js'
-export { default as URL } from '../modules/parseData/types/additional/URL.js'
+// `parseSheetData()`
+export { default as parseSheetData } from '../modules/parseSheetData/parseSheetData.js'
+export { default as Integer } from '../modules/parseSheetData/types/additional/Integer.js'
+export { default as Email } from '../modules/parseSheetData/types/additional/Email.js'
+export { default as URL } from '../modules/parseSheetData/types/additional/URL.js'